requirements.txt
gunicorn
flask
requests
redis
pyyaml
sqlalchemy
pymysql
python-dotenv
cryptography